+++ code/system.ui/react.monaco/src/ui.logic/EditorCarets/t.ts
